requirements.txt
slackclient == 1.3.2
python-dotenv==0.9.1
